refactor: clean missing schemas from datasets
