+++ src/argilla/server/schemas/v1/datasets.py
@@ -13,19 +13,12 @@
 #  limitations under the License.
 
 from datetime import datetime
-from typing import List, Literal, Optional, Union
+from typing import List, Optional
 from uuid import UUID
 
-import fastapi
-
-from argilla.server.enums import DatasetStatus, SimilarityOrder, SortOrder
-from argilla.server.pydantic_v1 import BaseModel, Field, constr, root_validator
+from argilla.server.enums import DatasetStatus
+from argilla.server.pydantic_v1 import BaseModel, Field, constr
 from argilla.server.schemas.base import UpdateSchema
-from argilla.server.schemas.v1.metadata_properties import MetadataPropertyName
-from argilla.server.schemas.v1.questions import QuestionName
-from argilla.server.schemas.v1.records import Record, RecordFilterScope
-from argilla.server.schemas.v1.responses import ResponseFilterScope
-from argilla.server.search_engine import TextQuery
 
 try:
     from typing import Annotated
@@ -38,15 +31,6 @@ DATASET_NAME_MAX_LENGTH = 200
 DATASET_GUIDELINES_MIN_LENGTH = 1
 DATASET_GUIDELINES_MAX_LENGTH = 10000
 
-TERMS_FILTER_VALUES_MIN_ITEMS = 1
-TERMS_FILTER_VALUES_MAX_ITEMS = 250
-
-FILTERS_AND_MIN_ITEMS = 1
-FILTERS_AND_MAX_ITEMS = 50
-
-SEARCH_RECORDS_QUERY_SORT_MIN_ITEMS = 1
-SEARCH_RECORDS_QUERY_SORT_MAX_ITEMS = 10
-
 
 DatasetName = Annotated[
     constr(regex=DATASET_NAME_REGEX, min_length=DATASET_NAME_MIN_LENGTH, max_length=DATASET_NAME_MAX_LENGTH),
@@ -58,6 +42,22 @@ DatasetGuidelines = Annotated[
     constr(min_length=DATASET_GUIDELINES_MIN_LENGTH, max_length=DATASET_GUIDELINES_MAX_LENGTH),
     Field(..., description="Dataset guidelines"),
 ]
+
+
+class RecordMetrics(BaseModel):
+    count: int
+
+
+class ResponseMetrics(BaseModel):
+    count: int
+    submitted: int
+    discarded: int
+    draft: int
+
+
+class DatasetMetrics(BaseModel):
+    records: RecordMetrics
+    responses: ResponseMetrics
 
 
 class Dataset(BaseModel):
@@ -92,145 +92,3 @@ class DatasetUpdate(UpdateSchema):
     allow_extra_metadata: Optional[bool]
 
     __non_nullable_fields__ = {"name", "allow_extra_metadata"}
-
-
-class RecordMetrics(BaseModel):
-    count: int
-
-
-class ResponseMetrics(BaseModel):
-    count: int
-    submitted: int
-    discarded: int
-    draft: int
-
-
-class DatasetMetrics(BaseModel):
-    records: RecordMetrics
-    responses: ResponseMetrics
-
-
-class MetadataParsedQueryParam:
-    def __init__(self, string: str):
-        k, *v = string.split(":", maxsplit=1)
-
-        self.name: str = k
-        self.value: str = "".join(v).strip()
-
-
-class MetadataQueryParams(BaseModel):
-    metadata: List[str] = Field(fastapi.Query([], pattern=r"^(?=.*[a-z0-9])[a-z0-9_-]+:(.+(,(.+))*)$"))
-
-    @property
-    def metadata_parsed(self) -> List[MetadataParsedQueryParam]:
-        # TODO: Validate metadata fields names from query params
-        return [MetadataParsedQueryParam(q) for q in self.metadata]
-
-
-class VectorQuery(BaseModel):
-    name: str
-    record_id: Optional[UUID] = None
-    value: Optional[List[float]] = None
-    order: SimilarityOrder = SimilarityOrder.most_similar
-
-    @root_validator(skip_on_failure=True)
-    def check_required(cls, values: dict) -> dict:
-        """Check that either 'record_id' or 'value' is provided"""
-        record_id = values.get("record_id")
-        value = values.get("value")
-
-        if bool(record_id) == bool(value):
-            raise ValueError("Either 'record_id' or 'value' must be provided")
-
-        return values
-
-
-class Query(BaseModel):
-    text: Optional[TextQuery] = None
-    vector: Optional[VectorQuery] = None
-
-
-class SuggestionFilterScope(BaseModel):
-    entity: Literal["suggestion"]
-    question: QuestionName
-    property: Optional[Union[Literal["value"], Literal["agent"], Literal["score"]]] = "value"
-
-
-class MetadataFilterScope(BaseModel):
-    entity: Literal["metadata"]
-    metadata_property: MetadataPropertyName
-
-
-FilterScope = Annotated[
-    Union[RecordFilterScope, ResponseFilterScope, SuggestionFilterScope, MetadataFilterScope],
-    Field(..., discriminator="entity"),
-]
-
-
-class TermsFilter(BaseModel):
-    type: Literal["terms"]
-    scope: FilterScope
-    values: List[str] = Field(..., min_items=TERMS_FILTER_VALUES_MIN_ITEMS, max_items=TERMS_FILTER_VALUES_MAX_ITEMS)
-
-
-class RangeFilter(BaseModel):
-    type: Literal["range"]
-    scope: FilterScope
-    ge: Optional[float]
-    le: Optional[float]
-
-    @root_validator(skip_on_failure=True)
-    def check_ge_and_le(cls, values: dict) -> dict:
-        ge, le = values.get("ge"), values.get("le")
-
-        if ge is None and le is None:
-            raise ValueError("At least one of 'ge' or 'le' must be provided")
-
-        if ge is not None and le is not None and ge > le:
-            raise ValueError("'ge' must have a value less than or equal to 'le'")
-
-        return values
-
-
-Filter = Annotated[Union[TermsFilter, RangeFilter], Field(..., discriminator="type")]
-
-
-class Filters(BaseModel):
-    and_: List[Filter] = Field(None, alias="and", min_items=FILTERS_AND_MIN_ITEMS, max_items=FILTERS_AND_MAX_ITEMS)
-
-
-class Order(BaseModel):
-    scope: FilterScope
-    order: SortOrder
-
-
-class SearchRecordsQuery(BaseModel):
-    query: Optional[Query]
-    filters: Optional[Filters]
-    sort: Optional[List[Order]] = Field(
-        None, min_items=SEARCH_RECORDS_QUERY_SORT_MIN_ITEMS, max_items=SEARCH_RECORDS_QUERY_SORT_MAX_ITEMS
-    )
-
-
-class SearchRecord(BaseModel):
-    record: Record
-    query_score: Optional[float]
-
-
-class SearchRecordsResult(BaseModel):
-    items: List[SearchRecord]
-    total: int = 0
-
-
-class SearchSuggestionOptionsQuestion(BaseModel):
-    id: UUID
-    name: str
-
-
-class SearchSuggestionOptions(BaseModel):
-    question: SearchSuggestionOptionsQuestion
-    agents: List[str]
-
-
-class SearchSuggestionsOptions(BaseModel):
-    items: List[SearchSuggestionOptions]
