+++ src/argilla/server/schemas/v1/suggestions.py
@@ -12,11 +12,12 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from typing import Any, List, Optional
+from typing import Any, List, Literal, Optional, Union
 from uuid import UUID
 
 from argilla.server.models import SuggestionType
 from argilla.server.pydantic_v1 import BaseModel, Field
+from argilla.server.schemas.v1.questions import QuestionName
 
 AGENT_REGEX = r"^(?=.*[a-zA-Z0-9])[a-zA-Z0-9-_:\.\/\s]+$"
 AGENT_MIN_LENGTH = 1
@@ -26,12 +27,43 @@ SCORE_GREATER_THAN_OR_EQUAL = 0
 SCORE_LESS_THAN_OR_EQUAL = 1
 
 
+class SuggestionFilterScope(BaseModel):
+    entity: Literal["suggestion"]
+    question: QuestionName
+    property: Optional[Union[Literal["value"], Literal["agent"], Literal["score"]]] = "value"
+
+
+class SearchSuggestionOptionsQuestion(BaseModel):
+    id: UUID
+    name: str
+
+
+class SearchSuggestionOptions(BaseModel):
+    question: SearchSuggestionOptionsQuestion
+    agents: List[str]
+
+
+class SearchSuggestionsOptions(BaseModel):
+    items: List[SearchSuggestionOptions]
+
+
 class BaseSuggestion(BaseModel):
     question_id: UUID
     type: Optional[SuggestionType]
     value: Any
     agent: Optional[str]
     score: Optional[float]
+
+
+class Suggestion(BaseSuggestion):
+    id: UUID
+
+    class Config:
+        orm_mode = True
+
+
+class Suggestions(BaseModel):
+    items: List[Suggestion]
 
 
 class SuggestionCreate(BaseSuggestion):
@@ -48,14 +80,3 @@ class SuggestionCreate(BaseSuggestion):
         le=SCORE_LESS_THAN_OR_EQUAL,
         description="The score assigned to the suggestion",
     )
-
-
-class Suggestion(BaseSuggestion):
-    id: UUID
-
-    class Config:
-        orm_mode = True
-
-
-class Suggestions(BaseModel):
-    items: List[Suggestion]
