+++ tests/unit/server/contexts/search/test_search_records_query_validator.py
@@ -17,7 +17,7 @@ from uuid import uuid4
 import argilla.server.errors.future as errors
 import pytest
 from argilla.server.contexts.search import SearchRecordsQueryValidator
-from argilla.server.schemas.v1.datasets import SearchRecordsQuery
+from argilla.server.schemas.v1.records import SearchRecordsQuery
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from tests.factories import (
